requirements.txt
Django>=6.0.0,<7.0.0
djangorestframework>=3.16.0,<4.0.0
django-filter>=25.0
django-simple-history>=3.11.0
psycopg2-binary>=2.9.0
asgiref>=3.9.1
sqlparse>=0.5.0
